chore: move app-header-item component into shared slice
