--- src/features/app-header/components/index.ts
@@ -1,1 +1,0 @@
-export * from './app-header-item'
