+++ src/shared/components/index.ts
@@ -1,3 +1,4 @@
 export * from './modal'
+export * from './nav-item'
 export * from './overlay'
 export * from './tabs'
