+++ src/shared/components/nav-item/style.module.css
@@ -1,25 +1,25 @@
-.itemButton {
+.item {
     padding: 1em 1.25em;
     display: flex;
     gap: 0.5em;
 }
 
-.itemButton:hover > svg {
+.item:hover > svg {
     opacity: 0.9;
     fill: var(--text-primary-color);
     transition: var(--common-transition);
 }
 
-.itemButton:not(:hover) > svg {
+.item:not(:hover) > svg {
     transition: var(--common-transition);
 }
 
-.itemButton:hover > span {
+.item:hover > span {
     color: var(--text-primary-color);
     transition: var(--common-transition);
 
 }
 
-.itemButton:not(:hover) > span {
+.item:not(:hover) > span {
     transition: var(--common-transition);
 }
