+++ src/shared/components/nav-item/type.ts
@@ -1,0 +1,11 @@
+import { FC, PropsWithChildren } from 'react'
+import { TIconProps } from 'uikit'
+
+interface Props {
+  to: string
+  title: string
+  size?: 'small' | 'medium' | 'large' | 'default'
+  Icon?: FC<TIconProps>
+}
+
+export type NavItemProps = PropsWithChildren<Props>
