+++ src/shared/components/nav-item/view.tsx
@@ -1,0 +1,28 @@
+import { useMatch, useNavigate } from 'react-router-dom'
+import { FC } from 'react'
+
+import { Button } from 'uikit'
+
+import { NavItemProps } from './type'
+import style from './style.module.css'
+
+export const NavItem: FC<NavItemProps> = ({ to, title, size = 'default', Icon }) => {
+  const nav = useNavigate()
+  const match = useMatch(to)
+
+  return (
+    <Button
+      onClick={() => nav(to)}
+      extraClass={style.item}
+      htmlType='button'
+      type='secondary'
+    >
+      {Icon && <Icon type={match ? 'primary' : 'secondary'}/>}
+      <span
+        className={`text text_type_main-${size} text_color_${match ? 'primary' : 'inactive'}`}
+      >
+        {title}
+      </span>
+    </Button>
+  )
+}
