+++ src/shared/uikit/index.ts
@@ -1,1 +1,4 @@
+import { TIconProps } from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/icons/utils'
+
 export * from '@ya.praktikum/react-developer-burger-ui-components'
+export type { TIconProps }
